# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fatourti Mobile Application APIS",
        "description": "API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/taxes": {
            "get": {
                "tags": [
                    "Taxes"
                ],
                "summary": "Get taxes for user",
                "description": "Get all tax records for the authenticated user's organization",
                "operationId": "getTaxesForUser",
                "responses": {
                    "200": {
                        "description": "Taxes retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "wording": {
                                                "type": "string",
                                                "example": "Tax Description"
                                            },
                                            "short_name": {
                                                "type": "string",
                                                "example": "Tax"
                                            },
                                            "value": {
                                                "type": "number",
                                                "example": 15.5
                                            },
                                            "value_type": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "application": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "organization_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Taxes"
                ],
                "summary": "Create a new tax record",
                "description": "Create a new tax record for the authenticated user's organization",
                "operationId": "createTax",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "wording": {
                                        "type": "string",
                                        "example": "Tax Description"
                                    },
                                    "short_name": {
                                        "type": "string",
                                        "example": "Tax"
                                    },
                                    "value": {
                                        "type": "number",
                                        "example": 15.5
                                    },
                                    "value_type": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "application": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tax created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "wording": {
                                            "type": "string",
                                            "example": "Tax Description"
                                        },
                                        "short_name": {
                                            "type": "string",
                                            "example": "Tax"
                                        },
                                        "value": {
                                            "type": "number",
                                            "example": 15.5
                                        },
                                        "value_type": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "application": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "organization_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/taxes/{id}": {
            "get": {
                "tags": [
                    "Taxes"
                ],
                "summary": "Get a specific tax",
                "description": "Get a specific tax record for the authenticated user's organization",
                "operationId": "getTax",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tax ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tax retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "wording": {
                                            "type": "string",
                                            "example": "Tax Description"
                                        },
                                        "short_name": {
                                            "type": "string",
                                            "example": "Tax"
                                        },
                                        "value": {
                                            "type": "number",
                                            "example": 15.5
                                        },
                                        "value_type": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "application": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "organization_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tax not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tax not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Taxes"
                ],
                "summary": "Update a tax",
                "description": "Update a specific tax record for the authenticated user's organization",
                "operationId": "updateTax",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tax ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "wording": {
                                        "type": "string",
                                        "example": "Updated Tax Description"
                                    },
                                    "short_name": {
                                        "type": "string",
                                        "example": "UpdTax"
                                    },
                                    "value": {
                                        "type": "number",
                                        "example": 20.5
                                    },
                                    "value_type": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "application": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tax updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "wording": {
                                            "type": "string",
                                            "example": "Updated Tax Description"
                                        },
                                        "short_name": {
                                            "type": "string",
                                            "example": "UpdTax"
                                        },
                                        "value": {
                                            "type": "number",
                                            "example": 20.5
                                        },
                                        "value_type": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "application": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "organization_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tax not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tax not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Taxes"
                ],
                "summary": "Delete a tax",
                "description": "Delete a specific tax record for the authenticated user's organization",
                "operationId": "deleteTax",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tax ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tax deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tax deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tax not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tax not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}